msm: mhi_dev: Reorganize the mhi header file
Reorganize the mhi header file. Expose MHI api's that are
required for clients.
