+++ drivers/platform/msm/mhi_dev/mhi_sm.h
@@ -13,7 +13,7 @@
 #ifndef MHI_SM_H
 #define MHI_SM_H
 
-#include <linux/msm_mhi_dev.h>
+#include "mhi.h"
 #include <linux/slab.h>
 #include <linux/msm_ep_pcie.h>
