+++ drivers/platform/msm/mhi_dev/mhi_uci.c
@@ -27,7 +27,7 @@
 #include <linux/msm_ipa.h>
 #include <linux/ipa.h>
 #include <uapi/linux/mhi.h>
-#include <linux/msm_mhi_dev.h>
+#include "mhi.h"
 
 #define MHI_DEV_NODE_NAME_LEN		13
 #define MHI_MAX_NR_OF_CLIENTS		23
